+++ examples/control/multi_agent.py
@@ -1,7 +1,9 @@
-import pathlib
-
 import gym
-from argument_parser import default_argument_parser
+import sys
+from pathlib import Path
+
+sys.path.insert(0, str(Path(__file__).parents[1]))
+from tools.argument_parser import default_argument_parser
 
 from smarts import sstudio
 from smarts.core.agent import Agent
@@ -65,7 +67,7 @@ if __name__ == "__main__":
     if not args.scenarios:
         args.scenarios = [
             str(
-                pathlib.Path(__file__).absolute().parents[1]
+                Path(__file__).absolute().parents[1]
                 / "scenarios"
                 / "sumo"
                 / "loop"
