+++ examples/egoless.py
@@ -1,7 +1,7 @@
 import logging
 
 import gym
-from argument_parser import default_argument_parser
+from tools.argument_parser import default_argument_parser
 
 from smarts.core.utils.episodes import episodes
