+++ examples/replay/replay_klws_agent.py
@@ -1,11 +1,13 @@
+import gym
 import logging
 import os
 import shutil
 import stat
 import sys
+from pathlib import Path
 
-import gym
-from argument_parser import default_argument_parser
+sys.path.insert(0, str(Path(__file__).parents[1]))
+from tools.argument_parser import default_argument_parser
 
 from smarts.core.utils.episodes import episodes
 from smarts.zoo.registry import make as zoo_make
