+++ examples/tests/test_examples.py
@@ -2,12 +2,11 @@ import sys
 import tempfile
 from pathlib import Path
 
-import psutil
 import pytest
 
 from smarts.core.utils import import_utils
 
-# necessary to import default_argument_parser properly in the examples
+# Necessary to import default_argument_parser in the examples
 sys.path.insert(0, str(Path(__file__).parents[1]))
 
 import_utils.import_module_from_file(
