+++ examples/trajectory_tracking_agent.py
@@ -1,12 +1,12 @@
 import logging
 
 import gym
-from argument_parser import default_argument_parser
 
 from smarts.core.agent import Agent
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.utils.episodes import episodes
 from smarts.zoo.agent_spec import AgentSpec
+from tools.argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
